refactor(AscensionSummary): move all material calculations in functions
