+++ src/utils/utils.tsx
@@ -15,6 +15,7 @@ import {characterExperience, HEROS_WIT_EXP as HEROS_WIT_EXP_QUANTITY} from '../c
 import CHARACTERS from '../charactersData';
 import {COUNT, MATERIAL, SORTING_INDEX} from '../const/consts';
 import {HEROS_WIT, MORA} from '../charactersData/materials/materialNames';
+import {IAscensionSummary} from '../views/Character/components/AscensionSummary/types';
 
 export const getOrgDataFromMatch = (matchData: match<IRouteParams>): string => {
   if (matchData?.params?.name) {
@@ -203,4 +204,52 @@ export const calculateMaterials = (characterName: string, startingLevel: ILevel,
   }
 
   return []
+}
+
+export const getCumulativeMaterials = ({ characterData, skillLevel, isTraveler}: IAscensionSummary): IMaterial[] => {
+  const skills = Object.keys(skillLevel)
+
+  const cumulativeMaterials: IMaterial[] = []
+
+  skills.forEach((skill: string) => {
+    const level = skillLevel[skill]
+
+    let talentMaterials: IAscensionMaterials = {}
+
+    if (isTraveler) talentMaterials = characterData[skill]
+
+    if (!isTraveler && characterData?.talent_materials) talentMaterials = characterData.talent_materials
+
+    return Object.keys(talentMaterials).forEach((skillLevel: string) => {
+      if (skillLevel > level.from && skillLevel <= level.to) {
+        return cumulativeMaterials.push(...talentMaterials[skillLevel])
+      }
+      return
+    })
+  })
+
+  return cumulativeMaterials
+}
+
+export const getGroupedMaterials = (materials: IMaterial[]): IMaterial[] => {
+  const materialGroupedByName = groupBy(materials, MATERIAL)
+  const groupedMaterialKeys = Object.keys(materialGroupedByName)
+
+  return groupedMaterialKeys.reduce((accumulator: IMaterial[], currentValue: string) => {
+    const counter: number = materialGroupedByName[currentValue].reduce((materialAccumulator: number, materialCurrentValue: IMaterial) => {
+      return materialAccumulator + materialCurrentValue.count
+    }, 0)
+
+    const material: IMaterial = {
+      material: currentValue,
+      count: counter,
+    }
+
+    const sortingIndex = materialGroupedByName[currentValue][0]?.sorting_index
+    if (sortingIndex) {
+      material[SORTING_INDEX] = sortingIndex
+    }
+
+    return [...accumulator, material]
+  }, [])
 }
